requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
openai==1.3.7
PyYAML==6.0.1
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
fake-useragent==1.4.0